+++ controllers/UpdateController.php
@@ -72,8 +72,16 @@ class UpdateController extends Controller
     {
         $availableUpdate = OnlineUpdateAPI::getAvailableUpdate();
 
+        /* @var \humhub\modules\marketplace\Module $marketplaceModule */
+        $marketplaceModule = Yii::$app->getModule('marketplace');
+        $updateModules = $marketplaceModule->onlineModuleManager->getAvailableUpdateModules();
+
         return $this->renderAjax('start', [
             'availableUpdate' => $availableUpdate,
+            'updateModules' => array_values(array_map(fn($updateModule) => [
+                'id' => $updateModule->id,
+                'name' => $updateModule->getName(),
+            ], $updateModules)),
         ]);
     }